requirements.txt
# Required Python packages for the Scrollhouse Onboarding Agent

# ── Core framework ─────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.7.1
httpx==0.27.0

# ── LangChain + LangGraph ──────────────────────────────────────────────────────
langchain==0.2.5
langchain-core==0.2.10
langchain-openai==0.1.8
langgraph==0.1.4

# ── LangSmith (optional but recommended) ──────────────────────────────────────
langsmith==0.1.77

# ── Email integration ──────────────────────────────────────────────────────────
sendgrid==6.11.0

# ── Google Drive ───────────────────────────────────────────────────────────────
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0

# ── Notion ─────────────────────────────────────────────────────────────────────
notion-client==2.2.1

# ── Airtable ───────────────────────────────────────────────────────────────────
pyairtable==2.3.3

# ── OpenAI (for AI-written emails) ────────────────────────────────────────────
openai==1.30.1
